refactor: 登録済み食品一覧の編集・削除モーダルをコンポーネント化（#12）
- useFoodApi.js: 食品データのAPI呼び出しとstate管理
- modalStyle.js: モーダルのスタイルを抽出
