+++ frontend/src/theme/modalStyle.js
@@ -1,0 +1,23 @@
+export const modalStyle = {
+  overlay: {
+    position: 'fixed',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+    backgroundColor: 'rgba(0, 0, 0, 0.3)'
+  },
+  content: {
+    position: "relative",
+    top: '50%',
+    left: '50%',
+    right: 'auto',
+    bottom: 'auto',
+    marginRight: '-50%',
+    transform: 'translate(-50%, -50%)',
+    maxWidth: "350px",
+    padding: "30px 40px",
+    borderRadius: "10px",
+    textAlign: "center",
+  }
+};
